backend/requirements.txt
fastapi
uvicorn[standard]
pystray
httpx
todoist-api-python
pillow
python-dotenv
google-genai
